error handling in lexer

+++ source/lib.h
@@ -1,9 +1,10 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stddef.h>
 
 int library_run_file(const char* filename);
 int library_run_prompt(void);
-void library_error(int line, const char* message);
+void library_error(size_t line, const char* message);
 
 extern bool HadError;
